pyproject.toml
[project]
name = "langgraph-examples"
version = "0.1.0"
description = "LangGraph 학습을 위한 통합 예제 패키지"
authors = [
    { name = "Isak" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
    # Core LangGraph
    "langgraph>=0.5.4",
    
    # LangChain Core
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.9",
    
    # LLM Providers
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.0",
    
    # Tools & Search
    "tavily-python>=0.5.0",
    
    # Utilities
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    
    # Data Processing (for multi_agent_supervisor)
    "defusedxml>=0.7.1",
    "langchain-experimental>=0.3.0",
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.1",
    "ruff>=0.6.1",
    "pytest>=8.0.0",
    "ipykernel>=6.29.5",
]

# Open Deep Research 전용 (별도 pyproject.toml 존재)
research = [
    "langchain-anthropic>=0.3.15",
    "langchain-mcp-adapters>=0.1.6",
    "langchain-deepseek>=0.1.2",
    "langchain-tavily",
    "langchain-groq>=0.2.4",
    "openai>=1.99.2",
    "arxiv>=2.1.3",
    "pymupdf>=1.25.3",
    "xmltodict>=0.14.2",
    "linkup-sdk>=0.2.3",
    "duckduckgo-search>=3.0.0",
    "exa-py>=1.8.8",
    "requests>=2.32.3",
    "beautifulsoup4>=4.13.3",
    "httpx>=0.24.0",
    "markdownify>=0.11.6",
    "rich>=13.0.0",
    "langgraph-cli[inmem]>=0.3.1",
    "langsmith>=0.3.37",
    "langchain-google-vertexai>=2.0.25",
    "azure-identity>=1.21.0",
    "azure-search>=1.0.0b2",
    "azure-search-documents>=11.5.2",
    "supabase>=2.15.3",
    "mcp>=1.9.4",
    "langchain-aws>=0.2.28",
    "pandas>=2.3.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
]
lint.ignore = [
    "E501", # line too long
]

[tool.ruff.lint.per-file-ignores]
"*/main.py" = ["E", "F"]
